#19: Repurpose SpringMover tests for LaplacianSmoother
